feat: Change strat from max to hash
