+++ teams/strategies_4.py
@@ -1,99 +1,138 @@
 import random
-from collections import defaultdict
+from itertools import chain
+from CardGame import Player, Deck, Card
 
-Pairs ={"Hearts": "Spades", "Spades": "Hearts", "Clubs": "Diamonds", "Diamonds": "Clubs"}
-PlayerPairs = {"North": "South", "South": "North", "East": "West", "West": "East"}
-max_suit_of_partner = None
-min_suit_of_partner = None
-max_suit_cards = ["2", "3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K", "A"]
-max_suit_of_partner1 = None
-min_suit_of_partner1 = None
-max_suit_cards1 = ["2", "3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K", "A"]
+random.seed(41)
 
+PARTNERS = {
+    "North": "South",
+    "East": "West",
+    "South": "North",
+    "West": "East",
+}
 
-def playing(player, deck):
-    """
-    Just return the opposite of maximum frequency suit. If opposite is not there, return min suit
-    """
-    if not player.hand:
-        return None
-    
-    freq = defaultdict(lambda:0)
-    
-    for card in player.hand:
-        freq[card.suit]+=1
-    
-    maxSuit =  max(freq, key=freq.get)
-    oppSuit = Pairs[maxSuit]
-    for i,card in enumerate(player.hand):
-        if card.suit == oppSuit:
-            return i
+SUIT_TO_NUM = {"Hearts": 0, "Spades": 13, "Diamonds": 26, "Clubs": 39}
 
-    # What if no opposite suit? Rn return min suit
-    minSuit = min(freq, key=freq.get)
-    for i,card in enumerate(player.hand):
-        if card.suit == minSuit:
-            return i
+VAL_TO_NUM = {
+    "A": 1,
+    "2": 2,
+    "3": 3,
+    "4": 4,
+    "5": 5,
+    "6": 6,
+    "7": 7,
+    "8": 8,
+    "9": 9,
+    "10": 10,
+    "J": 11,
+    "Q": 12,
+    "K": 13,
+}
+
+NUM_TO_SUIT = {v: k for k, v in SUIT_TO_NUM.items()}
+NUM_TO_VAL = {v: k for k, v in VAL_TO_NUM.items()}
 
 
-def guessing(player, cards, round):
-    global max_suit_of_partner1
-    global min_suit_of_partner1
-    global max_suit_cards1
-    global max_suit_of_partner
-    global min_suit_of_partner
-    global max_suit_cards
+def generate_permutation(perm_size):
+    """Generates a permutation dictionary, each card points to one permutation"""
+    numbers = list(range(1, 53))
+    perms = {}
 
-    exposedCards = player.exposed_cards[PlayerPairs[player.name]]
-    if player.name == "North" or player.name == "East":
-        max_suit_of_partner_local = max_suit_of_partner1
-        min_suit_of_partner_local = min_suit_of_partner1
-        max_suit_cards_local = max_suit_cards1
-    else:
-        max_suit_of_partner_local = max_suit_of_partner
-        min_suit_of_partner_local = min_suit_of_partner
-        max_suit_cards_local = max_suit_cards
+    for i in range(1, 53):
+        perms[i] = random.sample(numbers, perm_size)
+    return perms
 
-    if round == 1:
-        max_suit_of_partner_local = Pairs[exposedCards[0].suit]     
-    elif round == 2:
-        if player.cVals[-1] < 5: #Worst case 0H 5S 3C 4D 
-            # Suit guess was wrong. Why? Maybe opposite did not exist [No H]
-            min_suit_of_partner_local = max_suit_of_partner_local
-            all_suits = set(["Hearts", "Clubs", "Spades", "Diamonds"])
-            remove_suit = set([min_suit_of_partner_local, Pairs[min_suit_of_partner_local]])
-            max_suit_of_partner_local = random.choice(list(all_suits - remove_suit)) #Either H or S
-    elif round == 3:
-        if player.cVals[-2] < 5 and player.cVals[-1] < 5:  #0 5 2 4 
-            max_suit_of_partner_local = Pairs[max_suit_of_partner_local]
 
-    for p in player.exposed_cards:
-        for card in player.exposed_cards[p]:
-            if card.suit == max_suit_of_partner_local and card.value in max_suit_cards_local:
-                max_suit_cards_local.remove(card.value)
+def initialize_probabilities():
+    """Initilize probability dictionary"""
+    return {card: 1 / 52 for card in range(1, 53)}
 
-    # What if we have only 5 choices in max suit cards but we need to guess 6 cards?
-    if len(max_suit_cards_local) >= 13-round:
-        guesses = random.sample(max_suit_cards_local, 13-round) 
-    else:
-        guesses =  max_suit_cards_local 
-        guesses+= random.choices(max_suit_cards_local, k=13-round-len(max_suit_cards_local)) #Duplicates but we don't care
-    
-    # Right now set intersection in game gui does it with object ref not value so had to use cards passed - ineffecient 
-    returncards=[]
-    for guess in guesses:
-        for card in cards:
-            if card.suit == max_suit_of_partner_local and card.value == guess:
-                returncards.append(card)
-                break
 
-    if player.name == "North" or player.name == "East":
-        max_suit_of_partner1 = max_suit_of_partner_local
-        min_suit_of_partner1 = min_suit_of_partner_local
-        max_suit_cards1 = max_suit_cards_local
-    else:
-        max_suit_of_partner = max_suit_of_partner_local
-        min_suit_of_partner = min_suit_of_partner_local
-        max_suit_cards = max_suit_cards_local
+def initilize_permutations():
+    """Generates a dictionary with keys 1 to 12, values are permutations of decreasing size"""
+    result = {}
+    for key in range(1, 13):
+        perm_size = 13 - key
+        perms = generate_permutation(perm_size)
+        result[key] = perms
+    return result
 
-    return returncards
+
+PERMUTATIONS = initilize_permutations()
+
+
+def card_to_val(card: Card):
+    """Converts suit and value into a number from 1-52"""
+    return SUIT_TO_NUM[card.suit] + VAL_TO_NUM[card.value]
+
+
+def val_to_card(val: int) -> Card:
+    """Converts a number from 1-52 into a Card object"""
+    if val < 1 or val > 52:
+        raise ValueError("Value must be between 1 and 52")
+
+    for base in [39, 26, 13, 0]:  # Clubs, Diamonds, Spades, Hearts
+        if val > base:
+            suit = NUM_TO_SUIT[base]
+            break
+
+    # Determine the value
+    card_val = val - base
+    value = NUM_TO_VAL[card_val]
+
+    return Card(suit, value)
+
+
+def update_probabilities(
+    probs: dict[int, float],
+    guess: list[Card],
+    seen: list[Card],
+    correct: int,
+    hand_size: int,
+) -> dict[int, float]:
+    """Update probability dictionary"""
+    pass
+
+
+def playing(player: Player, deck: Deck):
+    game_round = len(player.played_cards) + 1
+    my_cards = [card_to_val(card) for card in player.hand]
+
+    # Find most similar permutations to players cards
+    card_index = 0
+    max_sim = 0
+    for i, k in enumerate(my_cards):
+        perms = PERMUTATIONS[game_round][k]
+        sim = len((set(my_cards) & set(perms)) - {k})
+        if sim > max_sim:
+            card_index = i
+            max_sim = sim
+
+    return card_index
+
+
+def guessing(player: Player, cards, round):
+
+    perm_index = card_to_val(player.exposed_cards[PARTNERS[player.name]][-1])
+
+    educated_guess = (
+        set(PERMUTATIONS[round][perm_index])
+        - set(card_to_val(card) for card in player.hand)
+        - set(chain.from_iterable(player.exposed_cards.values()))
+    )
+
+    guess_deck = list(
+        set(i for i in range(1, 53))
+        - set(card_to_val(card) for card in player.hand)
+        - set(
+            card_to_val(card)
+            for card in chain.from_iterable(player.exposed_cards.values())
+        )
+        - educated_guess
+    )
+
+    guess = random.sample(guess_deck, 13 - round - len(educated_guess))
+    educated_guess = list(educated_guess) + guess
+
+    card_guesses = [val_to_card(i) for i in educated_guess]
+    return card_guesses
